Feature #1757 added producer package, refactored tests and services according to new packages, added spring wolf autogeneration annotations
