+++ service/src/test/java/greencity/ubstelegrambot/TelegramSupportServiceTest.java
@@ -5,6 +5,7 @@ import greencity.dto.telegram.TelegramMessageDto;
 import greencity.entity.telegram.MessageAsset;
 import greencity.entity.telegram.TelegramChat;
 import greencity.entity.telegram.TelegramMessage;
+import greencity.producers.TelegramChatProducer;
 import greencity.repository.MessageAssetRepository;
 import greencity.repository.TelegramChatRepository;
 import greencity.repository.TelegramMessageRepository;
@@ -45,6 +46,9 @@ class TelegramSupportServiceTest {
 
     @Mock
     private TelegramNotificationService telegramNotificationService;
+
+    @Mock
+    private TelegramChatProducer telegramChatProducer;
 
     @Mock
     private TelegramMessageRepository telegramMessageRepository;
@@ -144,7 +148,7 @@ class TelegramSupportServiceTest {
 
         assertTrue(result.getText().contains(TelegramBotConstants.MESSAGE_SENT_TO_MANAGER_WAIT_FOR_RESPONSE));
         verify(telegramMessageRepository).save(any(TelegramMessage.class));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username, messageText, id);
     }
 
@@ -177,7 +181,7 @@ class TelegramSupportServiceTest {
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
         verify(telegramMessageRepository).save(any(TelegramMessage.class));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -215,7 +219,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService, never()).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -251,7 +255,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -292,7 +296,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService, never()).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -333,7 +337,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService, never()).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -369,7 +373,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -408,7 +412,7 @@ class TelegramSupportServiceTest {
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
         verify(telegramMessageRepository).save(any(TelegramMessage.class));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -451,7 +455,7 @@ class TelegramSupportServiceTest {
         SendMessage result = telegramSupportService.processSupportMessage(message);
 
         assertTrue(result.getText().contains(TelegramBotConstants.MANAGER_DIDNT_RECEIVED_YOUR_PHOTO_PLEASE_TRY_AGAIN));
-        verify(telegramNotificationService, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer, never()).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService, never()).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
     }
@@ -494,7 +498,7 @@ class TelegramSupportServiceTest {
 
         assertTrue(result.getText().contains(TelegramBotConstants.MESSAGE_SENT_TO_MANAGER_WAIT_FOR_RESPONSE));
         verify(telegramMessageRepository).save(any(TelegramMessage.class));
-        verify(telegramNotificationService).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
+        verify(telegramChatProducer).notifyNewMessage(any(TelegramMessageDto.class), anyLong());
         verify(telegramNotificationService).notifyManagerAboutNewMessagesFromUser(username,
             TelegramBotConstants.PHOTO_CONTENT, id);
         verify(fileService).upload(any(MultipartFile.class));
